backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
torch==2.3.1
torchvision==0.18.1
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
werkzeug==3.0.3
python-dotenv==1.0.1
